Tolerate cache folder deletion on writes.
If the cache is stored somewhere where the entire folder is subject
to possible deletion this will attempt to restore the folder tree
once on a `FileNotFoundException` during writes.

We choose not to perform the same operation on reads since no useful
state will be gained (i.e., it's a cache miss either way).

A test is included for the current ICS MR API. There is also a second
test (currently commented out) for the forthcoming public editing
API which Jesse Wilson was kind enough to make me privy to.

Change-Id: Ied6a3d8a7cc54eec7f923ef46b4d39a0599e221a