Don't end() user-supplied [Inf|Def]laters on closing stream
Calling close() on the following compression stream classes
*should not* invoke end() on the underlying Inflater/Deflater if it
was supplied in the stream class constructor:

    InflaterInputStream
    InflaterOutputStream
    DeflaterInputStream
    DeflaterOutputStream

This fix updates the behaviour of the close() method on those classes
to match the RI, allowing reuse of Inflater & Deflaters supplied to
stream class constructors.

This issue was initially observed using JGit on Android- JGit uses an
InflaterCache to recycle Inflaters, but in the cases where that Inflater
was used on a stream it would be returned to the cache in an invalidated
state. Seemingly-random exceptions would occur as the invalid Inflaters
were re-used in other contexts :-)

https://github.com/rtyley/agit/issues/47#issuecomment-3496045

Change-Id: Ia7b4b154820ba19ab08696978025c90b6fe2b1e5