Use default UncaughtExceptionHandler to report finalizer timeouts when defined.
In a previous release davlik would abort on finalizer timeouts which
would be noticed by various tools such as the monkey. However, more
recently when the handling of this was moved to managed code, we
simply logged and exited, which went unnoticed by the monkeys. By
reporting via the default UncaughtExceptionHandler, finalizer timeouts
will once again be noticed by the monkeys, this time via the
application crash reporting.

Bug: 6894375
Change-Id: I94664f8bb54dc7647b4d3666484d88b1c8428c4b