Copy changes to java.io.tmpdir to $TMPDIR.
This seems like the best way out to me. It means no new bionic API, it means
we work on glibc, it means native code now behaves the same as managed code,
and it means native code no longer ends up rewriting tmpfile(3) badly itself
because the built-in C library functions don't work (because they're hard-coded
to try /tmp and then give up).

There's currently a CTS test that explicitly checks that $TMPDIR isn't set,
so we'll have to disarm that.

Bug: 6938580
Bug: 7189929
Change-Id: If412ecda35f57eaa0b0a781b8dd36c851295cb4f