Secondary hash for IdentityHashMap and WeakHashMap.
Dalvik's identity hash code is a multiple of 8 meaning that it can cause
frequent collisions if a secondary hash isn't applied. This change adds
a secondary hash function to IdentityHashMap and WeakHashMap, as is done
in HashMap, HashTable and ConcurrentHashMap. There is some other minor clean-up
including updating the secondary hash function to match that in
ConcurrentHashMap.

Change-Id: Id2115b2622ff759bf2f67bd7572c2ecd8f0f0d67