HashMap: fix clone() warning
Fix IntelliJ warning:

This inspection reports clone() methods which are not declared as
throwing CloneNotSupportedException. If clone() is not declared to
possibly throw CloneNotSupportedException, then subclasses which
need to prohibit cloning will not be able to do so in the standard
way. This inspection will not report clone() methods declared final,
or clone() methods on final classes.

Change-Id: I13fd41e778f8af8c54060ed9753e128555ed5c06