Add implementation of AES-GCM
Adds implementation of AES in Galois/Counter Mode (GCM) using newer
APIs. This is the first Authenticated Encryption with Additional Data
(AEAD) mode supported. This means there is no need to use a MAC along
with the encryption to ensure it is not tampered with.

This is currently broken because Java implicitly adds the tag to the end
and we have no good way of getting it to supply it beforehand like
OpenSSL requires.

Change-Id: I3d288aeda74ad5e466bf79a4131857b4a2bf6c6d