CTS test fail because Zhuyin(Bopomofo) used.
The 'zh_TW' is a character set for Taiwanese language. Taiwanese
commonly uses Zhuyin(Bopomofo) for sorting their language
instead of Pinyin(for Chinese) or Stroke(for HK).But, the test
case tests the 'zh_TW' for Stroke, not for Zhuyin(Bopomofo).
For example, in the result, the expected character [7劃] is
a index label of Stoke, but the occurred character [ㄕ] is
that of Bopomofo

Signed-off-by: Raj Mamadgi <rmamadgi@sta.samsung.com>

(cherry picked from commit c5816949565963b0f6f1deb10413743db5c343f1)

Bug: 10349960
Change-Id: I3773bcb533be388ea7eb1ca7a961df87a0a0e3a2