Simplify code in XmlPullParserFactory.
Since we're hard coding KXmlParser and KXmlSerializer
anyway, we might as well do it more directly.

Also add a trivial unit test for this method.

NOTE: This change preserves existing behaviour and doesn't
throw from #newInstance if the parser class couldn't be
found or instantiated. We instead throw from newPullParser or
newSerializer as appropriate.

Change-Id: Ic0c86e67691b8e662d95363bbe597d3abb7c3642