Add VMRuntime.getTargetSdkVersion.
The goal is to enable libcore to adjust its behavior based on the
target SDK version / API Level of the app.

This change needs to be committed at the same time with the
supporting change in the dalvik project. This is because this change
renames a native method (setTargetSdkVersion) offered by the dalvik
project.

Change-Id: I63fa337a8e4175b64485cec432dd493111db8d4a