Make getInetAddress generate less garbage.
This function is now on the critical path for URLConnection
since OkHttp calls it to figure out what the MTU for the
NetworkInterface is.

This function performs poorly when there are a large
number of interfaces available.

- Read /proc/net/if_inet6 exactly once, instead of
  once per interface.
- Don't allocate an 8kb buffer to read 16 bytes from
  /sys/class/net/<ifname>/ifindex
- List the contents of /sys/class/net exactly once,
  and not once per interface.

Also adds a couple of tests for if_inet6 parsing

NOTE: We can make further improvements to getByInetAddress,
we could use the address family of the address to scan only
inet4 or inet6 addresses, and we break out of the interface
enumeration loop if we find a matching interface for the
address. We could also consider using rtnetlink / RTM_GETADDR
for AF_INET6 addresses if that makes sense.

We could also change OkHttp to cache the MTU for a given
interface name, if required.

bug: 11411129
Change-Id: Ieb1d5f0beaf462721f897a5f6376a4587f3aebaf