Late binding: Provider.Service#supportsParameter
First step in late-binding support. This will be used later to support
late binding by rejecting possible providers using the supportsParameter
call.

This allows security providers to specify attributes in their
registration that allow the Security framework to filter them out
earlier instead of failing on instantiation. The attributes supported in
this change are SupportedKeyClasses and SupportedKeyFormats.

Change-Id: Iee954e2b6c95c41d3b45aa738e869471fbe14341