Add missing JSONObject APIs.
- JSONObject JSONObject#put(String, java.util.Collection)

This is not backwards compatible. Before its introduction,
put("foo", (Collection) null) would remove the mapping for
key "foo". With this new API, put("foo", (Collection) null)
will clobber the existing mapping with a 0 length array.

The root cause for this madness is that put(String, null) is
specified to remove the mapping for "name". I'm not sure why
that was necessary, given the perfectly concise and usable
remove(String) method.

- Set JSONObject#keySet();

Also fixes a test failure in test_toString_listAsMapValue.

bug: https://code.google.com/p/android/issues/detail?id=63191
bug: https://code.google.com/p/android/issues/detail?id=63191
Change-Id: Idf2ef1889dccb4309c843ab073f67af648b09436