Removing the nestedException field from ObjectOutputStream.
The creation of the exception in the constructor adds expense.
A common usage of ObjectOutputStream in Android is to write Serializable
objects into a Parcel, which uses a ByteArrayOutputStream and should never fail
with an IOException. This change should provide a small speed up and reduce
memory usage for the common case.

Also tidying up other simple code problems:
1) Unused parameter on writeEnumDesc()
2) Removing the getFieldL native method. It is never called and does not appear
to be backed by any native code.
3) Removing unnecessary initialization of local variables.

bug: 6447733
Change-Id: If062aa8d376019ed0c8db969e8161f18898e64e5