Refactoring in preparation for NIO2 changes.
Making the various Socket/Channel classes more consistent with each
other for synchronizing state between the Channel, Socket and the OS. This is
in preparation for NetworkChannel.bind().
Synchronizing state revealed a test that relied upon a Socket from a channel
not reporting isConnected() properly after a connect(). Tests have been
updated.

Reading the local address back from the OS revealed that Android is using IPv6
and reports the string equivalent of IPv4's 0.0.0.0 as "::". Updated a test
that was relying on the string form.

Calling connect() twice on a DatagramSocket appears inconsistent with itself
and Channel and other Sockets. A Channel-backed DatagramSocket ignores the
second connect. A pure DatagramSocket switches over to the new remote address.
I have retained this behavior.

Socket has been modified to avoid using the address field to store the
local address after a bind(). This field is documented as holding the remote
address, not the local.

Changed implementation of SocketChannelImpl.socket().getInputStream() and
SocketChannelImpl.socket().getOutputStream() to use the one that is
implemented by Socket and not a custom NIO-based implementation. The use of
NIO provided two parallel implementations for the same thing. This change
alters behavior when attempting to read zero bytes and when a Channel is in
non-blocking mode. The tests have been updated.

Various tidying up changes to introduce @Override, remove unnecessarily
initialized local variables, fields, semi-colons, javadocs and imports.
Added close() calls for objects during tests that were cluttering the logs.

Change-Id: Ic7943615b4b763f77e74397e0e91a62edc7d7017
bug: 12464155