Fix ComputeSerialVersionUIDTest.
The old golden files were referring to the wrong class,
and contained serial version UIDs that didn't match any
known version of android (or the RI).

This change deletes the golden files and switches to
direct assertions in the test. All tests pass on
aosp-master, v4.xx of android and the RI.

bug: 12490445

Change-Id: Ia65b6d65bc9a3d4cbc62a7a0f7bdfc004e7315e2