Call chdir when "user.dir" is changed.
"user.dir" is the path used as the base for calculating
absolute paths. All java.io.File objects that don't specify
an absolute path are assumed to be relative to this path.

If we have bootstrap / setup code that changes this property
by calling System.setProperty, we should change our current
working directory accordingly. This allows code that uses
a mixture of getAbsolutePath and getPath on a file that was
constructed with a relative path to work correctly.

Note that I said "bootstrap" code because any app that changes
"user.dir" midway through execution is in for a bad time anyway.

bug: 12491159
bug: 12491204
Change-Id: Idd18f81a0bef5f12704a7d8231d4f5e7e0eb6c7a