Addition of 1.7 methods to URLConnection.
URLConnection.getHeaderLong() and
URLConnection.getContentLengthLong().
These methods are required by okhttp.

Changed JarURLConnectionImpl.getContentLength() to return -1
when the size of an entry cannot be represented as an int.
Previously it would have returned a corrupted, possibly
negative, value due to a cast.

Changed FileURLConnection.getContentLength() to return -1
when the size of the file cannot be represented as an int.
Previously it was returning Integer.MAX_VALUE.

Change-Id: Ib43e68a2536c2602b4c7ee0cda68fa1f90045f57