AArch64: LP64 fixes to libcore.
Make changes to native and Java code to support 64 bit or 32 bit native
pointers held in Java longs.

In particular, the MemoryTest class, in Java, used ints to do pointer
calculations. This was inadequate and was crashing on 64bit platforms.
The class was changed to use longs for pointer calculations.

Also, the startElement methods in the ExpatParser and EntityParser classes
was changed to use long rather than int for the attributePointer parameter.
The attributePointer attribute of the ExpatParser class was also changed
from int to long. The JNI use of the ExpatParser class's startElement
method was changed accordingly.

Change-Id: Idbc5b0d75fb1915e52f666f6078008d9e71f2c0b
Signed-off-by: Stuart Monteith <Stuart.Monteith@arm.com>
Signed-off-by: Marcus Oakland <marcus.oakland@arm.com>
Signed-off-by: Matteo Franchin <matteo.franchin@arm.com>
Signed-off-by: Kévin PETIT <kevin.petit@arm.com>