[WIP] Benchmark for HashMap, also experiment with open addressing.
This is completely unoptimized and fairly messy code. It does
seem correct and fully functional at this point, though.

Open addressing comes with several memory advantages.
- we don't need to allocate a new object for every put() method.
  we can operate the hash table with zero memory allocations (except
  for the table itself) unless entrySet() is called.
- as expected, the current benchmarks for put() score time higher
  for open addressing, but get() seems marginally slower. Methods that
  operate on values (containsValue, values) have become faster as well.

The intention here isn't to add any new API, but replace the
existing HashMap implementation if this new version is found to
be a better choice.

bug: 9111982

Change-Id: Ic02536fa26f89b0df714cd84038fe5fc7642d789