Explicate the contract of SSLContext.init with arrays.
SSLContext.init is supposed to treat null arrays of
KeyManager/TrustManager different from non-null arrays. In the case
of null arrays, SSLContext.init is supposed to find the highest
priority implementation of KeyManagerFactory/TrustManagerFactory and
then obtain the required types of KeyManager/TrustManager from the
factories. In the case of non-null arrays, this is not supposed to
happen. Instead, SSLContext.init is supposed to obtain each desired
type of KeyManager/TrustManager from the provided arrays.

Bug: 13563574
Bug: 13563675
Change-Id: I52adc5e7143d4f050be0b22b3b464c10bb97d102