Remove '.' from the System classpath for Android apps
When the Zygote is used, the java.class.path system property is
missing and so the System ClassLoader was being given the
default classpath of '.'. The meaning of '.' recently changed
from the app dir to '/'. Since any apps depending on this will
have broken with the '.' change, the System ClassLoader can be
changed to have an empty classpath, which is a more reasonable
choice.

Added documentation of this change inline. Fixed up some
long-standing commented code and removed a misleading JavaDoc
comment.

Change-Id: I64ac5a1edee118a6285df894d21f2ae3dc333408
Bug: 12491655