Change ReferenceQueue order from LIFO to FIFO.
LIFO order of FinalizerReference.queue causes incorrect work of runFinalization().
runFinalization() creates a Sentinel finalizable object and waits for it to be finalized.
As FinalizerReference.queue is "last in, first out" the Sentinel object is placed at the
beginning of the queue and then gets quickly finalized. So the method returns, although
"older" objects (placed in the queue before the Sentinel) were not finalized.
Changing FinalizerReference.queue to "first in, first out" fixes the issue
because finalizable objects are added to the end of the queue and "older"
ones will be already finalized when it comes to Sentinel.
Please also see issue 37980.

Change-Id: Ie6bc96e6a3759bc36ce6586066052b6d6bbe8d52
Signed-off-by: Elena Sayapina <elena.v.sayapina@intel.com>