Removing data hacks from ICU root.txt
Previously, the standard ICU data was modified
to support special short codes (e.g. PST, CET) in all
locales within SimpleDateFormat.parse().

This commit moves the modification into Java code
leaving the ICU data closer to upstream.

This change is associated with a change in external/icu4c:
I04acd15c62d49c0cf30cc63f60db320bdb8e22e9

Besides the SimpleDateFormat change, the removal
of the ICU data changes caused OldTimeZoneTest to
fail. It was relying on "PST" being a valid short
name for "America/Los_Angeles" for all locales.
The test was previously changed in 2012
(I503be2cebe8664a624de7b84f655b646db6e86f7): it previously
expected "UTC-08:00" and it was changed to "PST".
This looks like it was more to fit the data than
because it was specified behavior. Tests and comments
have been added to justify the difference with the RI
and prove that PST is not special in
TimeZone.getDisplayName().

Change-Id: I9ae6397cf5335ef325aedb6472d0d66a6127e1dd